% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corz.R
\name{corz}
\alias{corz}
\title{get correlation matrix with significance stars}
\usage{
corz(df, digits = 4)
}
\arguments{
\item{df}{dataframe for correlations}

\item{digits}{number of digits to print}
}
\value{
print of correlation matrix with sig stars
}
\description{
this function gets correlation matrix with sig stars
}
\examples{
\dontrun{
cor_matrix(df, digits = 4)
}
}

